example code for ContentStorage usage, text changes, codacy-forced refactoring
